requirements.txt
flask>=3.0.0
requests>=2.31.0
gunicorn>=21.0.0
google-auth>=2.29.0
google-api-python-client>=2.125.0
